@@ -46,17 +46,20 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
+    "dotenv": "^16.4.7",
     "embla-carousel-react": "^8.3.1",
     "framer-motion": "^11.12.0",
+    "he": "^1.2.0",
     "input-otp": "^1.4.1",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "lucide-react": "^0.456.0",
     "marked": "^15.0.2",
     "mssql": "^11.0.1",
-    "next": "15.0.3",
+    "next": "^15.1.0",
     "next-themes": "^0.4.3",
     "node-cache": "^5.1.2",
+    "pg": "^8.13.1",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-activity-calendar": "^2.7.1",
     "react-day-picker": "^8.10.1",
