@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
-    domains: ['corestore.blob.core.windows.net'],
+    domains: ['corestore.blob.core.windows.net', 'logo.clearbit.com'],
   },
 };
 
