@@ -1,5 +1,6 @@
 import { createDatabaseConnection } from "@/lib/db";
 import { getCached, setCached } from '@/lib/cache'; // ...existing code...
+import { query } from "@/lib/pgdb"; // Import the query method from db.js
 
 // /api/companies/job-postings-count/[id]
 // route to grab the amount of job postings for a company
@@ -7,23 +8,22 @@ import { getCached, setCached } from '@/lib/cache'; // ...existing code...
 export async function GET(req, { params }) {
 
     try {
-        console.log("Fetching job postings count for company:", params.id); 
-        const companyId = await params.id;
+        console.log("Fetching job postings count for company:", params.id);
+        const companyName = await params.id;
 
-        const cachedJobPostingsCount = getCached('job-postings-count', companyId);
+        const cachedJobPostingsCount = getCached('job-postings-count', companyName);
         if (cachedJobPostingsCount) {
             return new Response(JSON.stringify({ jobPostingsCount: cachedJobPostingsCount }), { status: 200 });
         }
-        const db = await createDatabaseConnection();
-        const result = await db.executeQuery(`
-            SELECT COUNT(*) AS jobPostingsCount
-            FROM jobPostings jp WITH (NOLOCK)
-            WHERE jp.company_id = @company_id;
-        `, { company_id: companyId });
-        
-        const jobPostingsCount = result.recordset[0].jobPostingsCount;
+        const result = await query(`
+            SELECT COUNT(*)
+            FROM jobPostings jp
+            WHERE jp.company = $1;
+        `, [companyName]);
 
-        setCached('job-postings-count', companyId, jobPostingsCount);
+        const jobPostingsCount = result.rows[0].count;
+
+        setCached('job-postings-count', companyName, jobPostingsCount);
 
         return new Response(JSON.stringify({ jobPostingsCount }), { status: 200 });
     } catch (error) {
