@@ -1,75 +1,43 @@
 import { NextResponse } from 'next/server';
 import jwt from 'jsonwebtoken';
-import { createDatabaseConnection } from '@/lib/db';
-import { getCompanies } from '@/lib/companyCache';
-import NodeCache from 'node-cache';
-const cache = new NodeCache({ stdTTL: 300 });
+import { query } from '@/lib/pgdb';
+const he = require('he');
 
 export async function GET(request) {
-  // Extract token from Authorization header
   const authHeader = request.headers.get('Authorization');
   if (!authHeader) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
   const token = authHeader.split(' ')[1];
 
-  if (!token) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const cachedUserJobs = cache.get(`userJobs:${token}`);
-  if (cachedUserJobs) {
-    return NextResponse.json({ userJobs: cachedUserJobs }, { status: 200 });
-  }
-  
   try {
-    console.log('JWT_SECRET:', process.env.SESSION_SECRET); // Debugging line
-    const decoded = jwt.verify(token, process.env.SESSION_SECRET || 'your-secret-key');
-    const userId = decoded.id; // Adjust based on token payload structure
-    console.log('Decoded userId:', decoded); // Debugging line
+    const decoded = jwt.verify(token, process.env.SESSION_SECRET);
+    const userId = decoded.id;
 
-    if (!userId) {
-      return NextResponse.json({ error: 'Invalid token payload' }, { status: 400 });
-    }
+    const result = await query(`
+      SELECT 
+        ui.interaction_date AS applied_at,
+        jp.*
+      FROM user_interactions ui
+      JOIN jobPostings jp ON ui.job_posting_id = jp.job_id
+      WHERE ui.user_id = $1 
+      AND ui.interaction_type = 'apply'
+      ORDER BY ui.interaction_date DESC
+    `, [userId]);
 
-    // Connect to the database
-    const [pool, companies ] = await Promise.all([createDatabaseConnection(), getCompanies()]);
-
-    // Query for user jobs from dbo.user_jobs
-    const query = `        
-          SELECT 
-          aj.id AS appliedJobId,
-          aj.applied_at AS appliedAt,
-          jp.id AS jobId,
-          jp.title as jobTitle, 
-          jp.location, 
-          jp.postedDate, 
-          jp.experienceLevel,
-          jp.salary
-        FROM user_jobs aj
-        INNER JOIN jobPostings jp ON aj.job_id = jp.id
-        WHERE aj.user_id = @userId
-        ORDER BY aj.applied_at DESC;`;
-    const userJobsResult = await pool.executeQuery(query, { userId });
-
-    const formattedUserJobs = userJobsResult.recordset.map(job => ({
-      id: job.jobId,
-      userId: job.userId,
-      favId: job.id,
-      postedDate: job.appliedAt,
-      status: job.status,
-      location: job.location,
-      salary: job.salary,
-      experienceLevel: job.experienceLevel,
-      jobStatus: job.jobStatus,
-      title: job.jobTitle,
-      company: job.companyName,
-      isCoreJob: job.isCoreJob,
+    const appliedJobs = result.rows.map(job => ({
+      id: job.job_id,
+      title: he.decode(job.title || ''),
+      company: job.company || '',
+      companyLogo: `https://logo.clearbit.com/${encodeURIComponent((job.company || '').replace('.com', ''))}.com`,
+      location: job.location || '',
+      experienceLevel: job.experiencelevel || '',
+      description: he.decode(job.description || ''),
+      postedDate: job.created_at?.toISOString() || '',
+      appliedAt: job.applied_at?.toISOString() || ''
     }));
 
-    cache.set(`userJobs:${token}`, formattedUserJobs);
-
-    return NextResponse.json({ userJobs: formattedUserJobs }, { status: 200 });
+    return NextResponse.json({ appliedJobs });
   } catch (error) {
     console.error('Error:', error);
     return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 });
