@@ -1,8 +1,14 @@
-
 import { NextResponse } from 'next/server';
-import { getConnection } from '@/lib/db';
+import { query } from '@/lib/pgdb';
 import jwt from 'jsonwebtoken';
 import sql from 'mssql';
+
+async function logInteraction(userId, jobId, interactionType) {
+  await query(`
+    INSERT INTO user_interactions (user_id, job_posting_id, interaction_type)
+    VALUES ($1, $2, $3)
+  `, [userId, jobId, interactionType]);
+}
 
 export async function POST(req, { params }) {
   try {
@@ -12,7 +18,7 @@ export async function POST(req, { params }) {
     if (!authHeader) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
-    
+
     const token = authHeader.split(' ')[1];
     const decoded = jwt.verify(token, process.env.SESSION_SECRET);
     const userId = decoded.id;
@@ -26,7 +32,7 @@ export async function POST(req, { params }) {
       await transaction.request()
         .input('jobId', sql.NVarChar, id)
         .query('UPDATE jobPostings SET views = ISNULL(views, 0) + 1 WHERE id = @jobId');
-      
+
       // Add to recently viewed
       await transaction.request()
         .input('userId', sql.NVarChar, userId)
@@ -43,6 +49,10 @@ export async function POST(req, { params }) {
         `);
 
       await transaction.commit();
+
+      // Log view interaction
+      await logInteraction(userId, id, 'view');
+
       return NextResponse.json({ success: true });
     } catch (error) {
       await transaction.rollback();
@@ -50,6 +60,50 @@ export async function POST(req, { params }) {
     }
   } catch (error) {
     console.error('Error tracking job view:', error);
+    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
+  }
+}
+
+async function bookmarkJob(req, { params }) {
+  try {
+    const { id } = await params;
+    const authHeader = req.headers.get('Authorization');
+    if (!authHeader) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
+    const token = authHeader.split(' ')[1];
+    const decoded = jwt.verify(token, process.env.SESSION_SECRET);
+    const userId = decoded.id;
+
+    // Log bookmark interaction
+    await logInteraction(userId, id, 'bookmark');
+
+    return NextResponse.json({ success: true });
+  } catch (error) {
+    console.error('Error bookmarking job:', error);
+    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
+  }
+}
+
+async function applyJob(req, { params }) {
+  try {
+    const { id } = await params;
+    const authHeader = req.headers.get('Authorization');
+    if (!authHeader) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
+    const token = authHeader.split(' ')[1];
+    const decoded = jwt.verify(token, process.env.SESSION_SECRET);
+    const userId = decoded.id;
+
+    // Log apply interaction
+    await logInteraction(userId, id, 'apply');
+
+    return NextResponse.json({ success: true });
+  } catch (error) {
+    console.error('Error applying to job:', error);
     return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
   }
 }
