@@ -1,86 +1,96 @@
-import { createDatabaseConnection } from "@/lib/db";
-import { getCompanies } from "@/lib/companyCache";
+import { query } from "@/lib/pgdb";
 import { getCached, setCached } from '@/lib/cache';
+const he = require('he');
 
-import sql from 'mssql';
+function scanKeywords(text) {
+  // ...existing code...
+}
 
-function formatForFullTextSearch(text) {
-  // Escape double quotes for SQL
-  return `"${text.replace(/"/g, '""')}"`;
+function extractSalary(text) {
+  // ...existing code...
 }
 
 export async function GET(req) {
   const { searchParams } = new URL(req.url);
 
-  // Extract and sanitize search filters
   const title = searchParams.get("title")?.trim() || "";
-  const experienceLevel = searchParams.get("experienceLevel")?.trim() || "";
-  const location = searchParams.get("location")?.trim() || "";
-  const company = searchParams.get("company")?.trim() || "";
-  const limit = parseInt(searchParams.get("limit")) || 10;
+  const experienceLevel = searchParams.get("experienceLevel")?.trim().toLowerCase() || "";
+  const location = searchParams.get("location")?.trim().toLowerCase() || "";
+  const limit = parseInt(searchParams.get("limit")) || 5;
 
-  const cachedJobs = getCached('jobs-by-saved-search', { title, experienceLevel, location, company, limit });
+  const cacheKey = `jobs-by-saved-search-${title}-${experienceLevel}-${location}-${limit}`;
+  const cachedJobs = getCached(cacheKey);
   if (cachedJobs) {
     return new Response(JSON.stringify({ jobs: cachedJobs }), { status: 200 });
   }
 
   try {
-    const [pool, companies] = await Promise.all([createDatabaseConnection(), getCompanies()]);
-
-    // Build the query to fetch job postings
-    let query = `
+    let queryText = `
       SELECT 
-        jp.id, jp.title, jp.experienceLevel, jp.location, jp.company_id, jp.postedDate
-      FROM jobPostings jp WITH (NOLOCK)
-      WHERE jp.deleted = 0
+        id, 
+        job_id,
+        source_url,
+        experiencelevel,
+        title, 
+        company, 
+        location,
+        description,
+        created_at
+      FROM jobPostings
+      WHERE 1 = 1
     `;
 
-    if (title) query += ` AND FREETEXT(jp.title, @title)`;
-    if (experienceLevel) query += ` AND jp.experienceLevel = @experienceLevel`;
-    if (location) query += ` AND CONTAINS(jp.location, @location)`;
-    if (company) query += ` AND jp.company_id = @company_id`;
+    const params = [];
+    let paramIndex = 1;
 
-    query += ` ORDER BY jp.postedDate DESC OFFSET 0 ROWS FETCH NEXT @limit ROWS ONLY`;
-
-    // Consolidate parameters
-    let params = {
-      limit
-    };
-
-    if (title) params.title = formatForFullTextSearch(title);
-    if (experienceLevel) params.experienceLevel = experienceLevel;
-    if (location) params.location = formatForFullTextSearch(location);
-    if (company) {
-      // If filtering by company, map company name to company_id
-      const companyResult = await pool.executeQuery(`
-        SELECT id FROM companies WITH (NOLOCK)
-        WHERE LOWER(name) = LOWER(@companyName) AND deleted = 0
-      `, { companyName: company });
-
-      if (companyResult.recordset.length === 0) {
-        // No matching company found, return empty result
-        return new Response(JSON.stringify({ jobs: [] }), { status: 200 });
-      }
-
-      const companyId = companyResult.recordset[0].id;
-      params.company_id = companyId;
+    if (title) {
+      queryText += ` AND title ILIKE $${paramIndex}`;
+      params.push(`%${title}%`);
+      paramIndex++;
     }
 
-    const result = await pool.executeQuery(query, params);
-    let jobs = result.recordset;
+    if (experienceLevel) {
+      queryText += ` AND experiencelevel ILIKE $${paramIndex}`;
+      params.push(`%${experienceLevel}%`);
+      paramIndex++;
+    }
 
-    // Add company name to each job posting
-    jobs = jobs.map(job => {
-      const company = companies[job.company_id];
+    if (location) {
+      queryText += ` AND location ILIKE $${paramIndex}`;
+      params.push(`%${location}%`);
+      paramIndex++;
+    }
+
+    queryText += `
+      ORDER BY created_at DESC
+      LIMIT $${paramIndex}
+    `;
+    params.push(limit);
+
+    const result = await query(queryText, params);
+
+    const jobs = result.rows.map((job) => {
+      const keywords = scanKeywords(job.description);
+      const remoteKeyword = job.location.toLowerCase().includes('remote') ? 'Remote' : "";
+      const salary = extractSalary(job.description);
+
       return {
-        ...job,
-        companyName: company ? company.name : "Unknown",
-        companyLogo: company ? company.logo : null,
+        id: job.job_id,
+        title: job.title || "",
+        company: job.company || "",
+        companyLogo: `https://logo.clearbit.com/${encodeURIComponent(job.company.replace('.com', ''))}.com`,
+        experienceLevel: job.experiencelevel || "",
+        description: he.decode(job.description || ""),
+        location: job.location || "",
+        salary: salary,
+        postedDate: job.created_at ? job.created_at.toISOString() : "",
+        remoteKeyword: remoteKeyword,
+        keywords: keywords,
       };
     });
 
-    setCached('jobs-by-saved-search', { title, experienceLevel, location, company, limit }, jobs);
-    
+    setCached(cacheKey, jobs);
+
     return new Response(JSON.stringify({ jobs }), { status: 200 });
   } catch (error) {
     console.error("Error fetching jobs by saved search:", error);
