@@ -1,60 +1,48 @@
-
-import { getConnection } from "@/lib/db";
-import sql from 'mssql';
+import { query } from "@/lib/pgdb";
 import { getCompanies } from "@/lib/companyCache";
 
 export async function GET(req) {
   const { searchParams } = new URL(req.url);
-  const companyId = searchParams.get("company_id");
+  const company = searchParams.get("company");
   const limit = parseInt(searchParams.get("limit")) || 3;
 
-  if (!companyId) {
-    return new Response(JSON.stringify({ error: "Job ID required" }), { status: 400 });
+  if (!company) {
+    return new Response(JSON.stringify({ error: "Company required" }), { status: 400 });
   }
 
   try {
-    const pool = await getConnection();
-    const companies = await getCompanies();
-    
-    const request = pool.request()
-      .input('id', sql.Int, parseInt(companyId))
-      .input('limit', sql.Int, limit);
 
-    const query = `
-      WITH SourceJob AS (
-        SELECT title, experienceLevel, location
-        FROM jobPostings WITH (NOLOCK)
-        WHERE id = @id
-      )
-      SELECT TOP (@limit)
-        jp.id, jp.title, jp.location, jp.postedDate, 
-        jp.salary, jp.experienceLevel, jp.company_id
-      FROM jobPostings jp WITH (NOLOCK)
-      CROSS JOIN SourceJob sj
-      WHERE jp.company_id != @id 
-        AND jp.deleted = 0
-        AND (
-          jp.experienceLevel = sj.experienceLevel
-          OR jp.location LIKE '%' + sj.location + '%'
-        )
-      ORDER BY jp.postedDate DESC`;
+    const similarQuery = `
+      SELECT 
+        job_id as id,
+        title,
+        location,
+        created_at as "postedDate",
+        company,
+        salary,
+        experiencelevel,
+      FROM jobpostings
+      WHERE company = $1 
+        AND deleted = false
+      ORDER BY created_at DESC
+      LIMIT $2`;
 
-    const result = await request.query(query);
+    const result = await query(similarQuery, [company, limit]);
 
-    const similarJobs = result.recordset.map(job => ({
+    const similarJobs = result.rows.map(job => ({
       id: job.id,
       title: job.title,
-      company: companies[job.company_id]?.name || "Unknown",
-      experienceLevel: job.experienceLevel,
+      company: job.company || "Unknown",
+      experienceLevel: job.experiencelevel,
       location: job.location,
       salary: job.salary,
-      logo: companies[job.company_id]?.logo || null,
+      logo: 'null',
       postedDate: job.postedDate
     }));
 
     return new Response(JSON.stringify({ similarJobs }), { status: 200 });
   } catch (error) {
-    console.error("Error fetching similar jobs:", error);
-    return new Response(JSON.stringify({ error: "Error fetching similar jobs" }), { status: 500 });
+    console.error("Error fetching similar company jobs:", error);
+    return new Response(JSON.stringify({ error: "Error fetching similar company jobs" }), { status: 500 });
   }
 }
