@@ -6,10 +6,10 @@ import { formatDistanceToNow } from "date-fns";
 import AlertDemo from "./AlertDemo";
 import { useAuth } from '@/context/AuthContext';
 import {
-    Accordion,
-    AccordionContent,
-    AccordionItem,
-    AccordionTrigger,
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
 } from "@/components/ui/accordion"
 import {
   Breadcrumb,
@@ -39,13 +39,13 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover"
-import { ArrowRight, Briefcase, Flag, Mail, MapPin, Sparkle, Timer, User, Wand2, Zap, DollarSign, Sparkles } from "lucide-react";
+import { ArrowRight, Briefcase, Bell, Flag, Mail, MapPin, Sparkle, Timer, User, Wand2, Zap, DollarSign, Sparkles } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
 import { JobCard } from "../../../components/job-posting";
 import { CollapsibleDemo } from "./collapsible";
 import { StickyNavbar } from './navbar';
 const stripHTML = (str) => {
-  const allowedTags = ['b', 'i', 'strong', 'em', 'u', 'ul', 'ol', 'li', 'p', 'br', 'span', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'div'];
+  const allowedTags = ['b', 'i', 'strong', 'br', 'em', 'u'];
   const parser = new DOMParser();
   const doc = parser.parseFromString(str, 'text/html');
 
@@ -78,7 +78,7 @@ import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/comp
 import { cn } from "@/lib/utils";
 import { Check, Copy } from "lucide-react";
 
-export function CopyButton() {
+function CopyButton() {
   const [copied, setCopied] = useState(false);
   const [currentUrl, setCurrentUrl] = useState("");
 
@@ -134,17 +134,48 @@ export function CopyButton() {
   );
 }
 
+function MagicButton() {
+  return (
+    <Popover>
+      <PopoverTrigger asChild>
+        <Button
+          variant="outline"
+          size="icon"
+          className="disabled:opacity-100 group"
+        >
+          <Sparkles className="group-hover:stroke-emerald-500" size={16} strokeWidth={2} aria-hidden="true" />
+        </Button>
+      </PopoverTrigger>
+      <PopoverContent className="max-w-[280px] py-3 ml-4 shadow-none" side="top">
+        <div className="space-y-3">
+          <div className="space-y-1">
+            <p className="text-[13px] font-medium">Generate Summary</p>
+            <p className="text-xs text-muted-foreground">
+              Instantly generate a summary of the job posting using AI models.
+            </p>
+          </div>
+          <Button size="sm" className="h-7 px-2">
+            Generate Summary
+          </Button>
+        </div>
+      </PopoverContent>
+    </Popover>
+  );
+}
+
+
+
 // Create a separate component for Similar Jobs
-const SimilarJobs = ({ jobId }) => {
+const SimilarJobs = ({ jobTitle, experienceLevel }) => {
   const [similarJobs, setSimilarJobs] = useState([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     const fetchSimilarJobs = async () => {
       try {
-        const response = await fetch(`/api/job-postings/similar?id=${jobId}`);
+        const response = await fetch(`/api/job-postings?title=${jobTitle}&experienceLevel=${experienceLevel}`);
         const data = await response.json();
-        setSimilarJobs(data.similarJobs);
+        setSimilarJobs(data.jobPostings);
       } catch (error) {
         console.error('Error fetching similar jobs:', error);
       } finally {
@@ -153,29 +184,29 @@ const SimilarJobs = ({ jobId }) => {
     };
 
     fetchSimilarJobs();
-  }, [jobId]);
+  }, [jobTitle, experienceLevel]);
 
   if (loading) return <div>Loading similar jobs...</div>;
-  
+
   return (
-    <CollapsibleDemo 
-      title="Similar Job Postings" 
+    <CollapsibleDemo
+      title="Similar Job Postings"
       open={true}
-      jobPostings={similarJobs} 
+      jobPostings={similarJobs}
     />
   );
 };
 
-const CompanySimilarJobs = ({ companyId }) => {
+const CompanySimilarJobs = ({ company }) => {
   const [similarJobs, setSimilarJobs] = useState([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     const fetchSimilarJobs = async () => {
       try {
-        const response = await fetch(`/api/job-postings/similar-company?company_id=${companyId}`);
+        const response = await fetch(`/api/job-postings/?company=${company}`);
         const data = await response.json();
-        setSimilarJobs(data.similarJobs);
+        setSimilarJobs(data.jobPostings);
       } catch (error) {
         console.error('Error fetching similar jobs:', error);
       } finally {
@@ -184,7 +215,7 @@ const CompanySimilarJobs = ({ companyId }) => {
     };
 
     fetchSimilarJobs();
-  }, [companyId]);
+  }, [company]);
 
   if (loading) return <div>Loading similar jobs...</div>;
 
@@ -199,7 +230,7 @@ const CompanySimilarJobs = ({ companyId }) => {
 
 
 
-export function InsightsButton({ onClick }) {
+function InsightsButton({ onClick }) {
   return (
     <Button variant="outline" onClick={onClick}>
       Show Insights
@@ -208,13 +239,137 @@ export function InsightsButton({ onClick }) {
   );
 }
 
+// ### Updated Utility Function: Extract Salary ###
+/**
+ * Extracts salary information from a given text.
+ * Handles various formats such as:
+ * - 'USD $100,000-$200,000'
+ * - '$100k-120k'
+ * - '55/hr - 65/hr'
+ * - '$4200 monthly'
+ * - etc.
+ *
+ * @param {string} text - The text to extract salary from.
+ * @returns {string} - The extracted salary string or an empty string if not found.
+ */
+function extractSalary(text) {
+  if (!text) return "";
+
+  // Step 1: Decode HTML entities
+  // Create a temporary DOM element to leverage the browser's HTML parser
+  const decodedString = he.decode(text);
+
+  // Step 2: Remove HTML tags
+  const textWithoutTags = decodedString.replace(/<[^>]*>/g, ' ');
+
+  // Step 3: Normalize HTML entities and special characters
+  const normalizedText = textWithoutTags
+    .replace(/\u00a0/g, ' ')       // Replace non-breaking spaces
+    .replace(/&nbsp;/g, ' ')       // Replace &nbsp;
+    .replace(/&mdash;/g, '—')      // Replace &mdash; with em-dash
+    .replace(/&amp;/g, '&')        // Replace &amp; with &
+    .replace(/&lt;/g, '<')         // Replace &lt; with <
+    .replace(/&gt;/g, '>')         // Replace &gt; with >
+    .trim();
+
+  // Define regex patterns
+  const patterns = [
+    // 1. Salary ranges with dashes (e.g., "$128,000—$152,000 USD")
+    /\$\s*(\d{1,3}(?:,\d{3})+|\d{3,})\s*[-–—]\s*\$\s*(\d{1,3}(?:,\d{3})+|\d{3,})\s*(USD|CAD)?(?:\s*per\s*year)?/gi,
+
+    // 2. Salary ranges with 'to' wording (e.g., "$35,000 to $45,000 per year")
+    /\$\s*(\d{1,3}(?:,\d{3})+|\d{3,})\s*(to|through|up\s*to)\s*\$\s*(\d{1,3}(?:,\d{3})+|\d{3,})\s*(USD|CAD)?(?:\s*per\s*year)?/gi,
+
+    // 3. k-based salary ranges (e.g., "$100k—$120k")
+    /\$\s*(\d+\.?\d*)k\s*[-–—]\s*\$\s*(\d+\.?\d*)k/gi,
+
+    // 4. Hourly ranges (e.g., "55/hr - 65/hr")
+    /(\d+\.?\d*)\s*[-–—]\s*(\d+\.?\d*)\s*\/\s*(hour|hr|h)/gi,
+
+    // 5. Monthly salaries with at least three digits (e.g., "$4200 monthly")
+    /\$\s*(\d{3,}\.?\d*)\s*\b(monthly|month|months|mo)\b/gi,
+
+    // 6. Single salary mentions (e.g., "$85,000")
+    /\$\s*\d{1,3}(?:,\d{3})+(?:\.\d+)?\b/gi,
+  ];
+
+  let matchesWithDollar = [];
+  let matchesWithoutDollar = [];
+
+  // Iterate through each pattern and collect matches
+  for (const pattern of patterns) {
+    const matches = Array.from(normalizedText.matchAll(pattern));
+    for (const match of matches) {
+      if (pattern.source.includes('\\$')) {
+        // Patterns that require '$' are stored in matchesWithDollar
+        matchesWithDollar.push({
+          text: match[0].trim(),
+          index: match.index
+        });
+      } else {
+        // Patterns that do NOT require '$' are stored in matchesWithoutDollar
+        matchesWithoutDollar.push({
+          text: match[0].trim(),
+          index: match.index
+        });
+      }
+    }
+  }
+
+  // Function to find the match with the highest index
+  const getLastMatch = (matches) => {
+    return matches.reduce((prev, current) => {
+      return (prev.index > current.index) ? prev : current;
+    }, matches[0]);
+  };
+
+  // Prioritize matches with '$'
+  if (matchesWithDollar.length > 0) {
+    const lastMatch = getLastMatch(matchesWithDollar);
+    return lastMatch.text;
+  }
+  // If no matches with '$', consider matches without '$'
+  else if (matchesWithoutDollar.length > 0) {
+    const lastMatch = getLastMatch(matchesWithoutDollar);
+    return lastMatch.text;
+  }
+
+  // Return empty string if no matches found
+  return "";
+}
+
+export function BellButton({ count }) {
+
+  const handleClick = () => {
+    setCount(0);
+  };
+
+  return (
+    <Button
+      variant="outline"
+      size="icon"
+      className="relative"
+      onClick={handleClick}
+      aria-label="Notifications"
+    >
+      <Bell size={16} strokeWidth={2} aria-hidden="true" />
+      {count > 0 && (
+        <Badge className="absolute -top-2 left-full min-w-5 -translate-x-1/2 px-1">
+          {count}
+        </Badge>
+      )}
+    </Button>
+  );
+}
+
+
 export default function JobPostingPage({ params }) {
   const { id } = use(params);
   const [data, setData] = useState(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const [insightsShown, setInsightsShown] = useState(false);
-  const { user } = useAuth(); 
+  const { user } = useAuth();
   const [showAlert, setShowAlert] = useState(false);
   const [llmResponse, setLlmResponse] = useState("");
   const [userProfile, setUserProfile] = useState(null);
@@ -225,7 +380,7 @@ export default function JobPostingPage({ params }) {
 
   const handleBadgeClick = () => {
     setShowAlert(true);
-};
+  };
 
   const handlePredefinedQuestion = async (e) => {
     if (!user) return;
@@ -243,7 +398,7 @@ export default function JobPostingPage({ params }) {
     const softSkills = userProfile.user.soft_skills || 'None specified';
     const otherSkills = userProfile.user.other_skills || 'None specified';
 
-    const modifiedQuestion = `Does ${userProfile.user.firstname} qualify for the job titled "${jobPosting.title}" at ${jobPosting.companyName}? Please provide a match score out of 100 and a brief explanation.`;
+    const modifiedQuestion = `Does ${userProfile.user.firstname} qualify for the job titled "${jobPosting.title}" at ${jobPosting.company}? Please provide a match score out of 100 and a brief explanation.`;
     const matchSchema = {
       type: "json_schema",
       json_schema: {
@@ -277,22 +432,22 @@ You are a helpful career assistant evaluating job fit for ${userProfile.user.fir
 - **Willing to Relocate:** ${userProfile.user.willing_to_relocate ? 'Yes' : 'No'}
 
 ### Work Experience
-${userProfile.experience && userProfile.experience.length > 0 
-? userProfile.experience.map(exp => 
-    `- **${exp.title}** at **${exp.companyName}** (${new Date(exp.startDate).toLocaleDateString()} - ${exp.isCurrent ? 'Present' : new Date(exp.endDate).toLocaleDateString()})
+${userProfile.experience && userProfile.experience.length > 0
+          ? userProfile.experience.map(exp =>
+            `- **${exp.title}** at **${exp.company}** (${new Date(exp.startDate).toLocaleDateString()} - ${exp.isCurrent ? 'Present' : new Date(exp.endDate).toLocaleDateString()})
 - **Location**: ${exp.location || 'Not specified'}
 - **Description**: ${exp.description || 'No description available'}
 - **Tags**: ${exp.tags || 'No tags available'}`).join('\n\n')
-: 'No work experience available.'}
+          : 'No work experience available.'}
 
 ### Education
 ${userProfile.education && userProfile.education.length > 0
-? userProfile.education.map(edu => 
-    `- **${edu.degree} in ${edu.fieldOfStudy}** from **${edu.institutionName}**
+          ? userProfile.education.map(edu =>
+            `- **${edu.degree} in ${edu.fieldOfStudy}** from **${edu.institutionName}**
 - **Duration**: ${new Date(edu.startDate).toLocaleDateString()} - ${edu.isCurrent ? 'Present' : new Date(edu.endDate).toLocaleDateString()}
 - **Grade**: ${edu.grade || 'Not specified'}
 - **Activities**: ${edu.activities || 'No activities specified'}`).join('\n\n')
-: 'No education details available.'}
+          : 'No education details available.'}
 
 ### Job Posting Details:
 ${JSON.stringify(jobPosting)}
@@ -348,9 +503,9 @@ Please assess the qualifications and provide a brief explanation of whether the 
       }
     }
 
-    async function fetchCompanyJobCount(id) {
+    async function fetchCompanyJobCount(companyName) {
       try {
-        const response = await fetch(`/api/companies/job-postings-count/${id}`, {
+        const response = await fetch(`/api/companies/job-postings-count/${companyName}`, {
           signal: controller.signal
         });
         console.log('Company job count response:', response);
@@ -372,9 +527,9 @@ Please assess the qualifications and provide a brief explanation of whether the 
       try {
         const token = localStorage.getItem('token');
         const headers = token ? { Authorization: `Bearer ${token}` } : {};
-        
+
         // Fetch job data
-        const response = await fetch(`/api/job-postings/${id}`, { 
+        const response = await fetch(`/api/job-postings/${id}`, {
           headers,
           signal: controller.signal
         });
@@ -382,9 +537,9 @@ Please assess the qualifications and provide a brief explanation of whether the 
         const result = await response.json();
         if (isMounted) {
           setData(result);
-          fetchCompanyJobCount(result.data.company_id);
           setLoading(false);
-          
+          fetchCompanyJobCount(result.data.company);
+
           // Only track view if we successfully fetched the job data and have a token
           if (token && !sessionStorage.getItem(`viewed-${id}`)) {
             try {
@@ -422,6 +577,23 @@ Please assess the qualifications and provide a brief explanation of whether the 
     };
   }, [id, user]); // Only depend on id and user
 
+  const handleApplyClick = async () => {
+    if (!user) return; // Only track if user is logged in
+
+    try {
+      await fetch('/api/applications', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${user.token}`,
+        },
+        body: JSON.stringify({ jobPostingId: id }),
+      });
+    } catch (error) {
+      console.error('Error tracking application:', error);
+    }
+  };
+
   if (loading) return <div>Loading...</div>;
   if (error) return <div>Error: {error}</div>;
   if (!data.success) return <div>Job posting not found.</div>;
@@ -431,13 +603,13 @@ Please assess the qualifications and provide a brief explanation of whether the 
 
   const { keywords, relatedPostings } = data;
 
-  return (  
+  return (
     <div className="container mx-auto py-6 px-4 sm:px-6 lg:px-8 max-w-4xl">
-      <StickyNavbar 
+      <StickyNavbar
         title={jobPosting.title}
-        companyName={jobPosting.companyName}
-        companyLogo={jobPosting.logo}
-        companyId={jobPosting.company_id}
+        companyName={jobPosting.company}
+        companyLogo={''}
+        companyId={1}
       />
       <Breadcrumb className="mb-4">
         <BreadcrumbList>
@@ -446,247 +618,244 @@ Please assess the qualifications and provide a brief explanation of whether the 
           </BreadcrumbItem>
           <BreadcrumbSeparator />
           <BreadcrumbItem>
-            <BreadcrumbLink href={`/companies/${jobPosting.company_id}`}>{jobPosting.companyName}</BreadcrumbLink>
+            <BreadcrumbLink href={`/job-postings?company=${jobPosting.company}`}>{jobPosting.company}</BreadcrumbLink>
           </BreadcrumbItem>
           <BreadcrumbSeparator />
           <BreadcrumbItem>
-            <BreadcrumbPage>{jobPosting.id}</BreadcrumbPage>
+            <BreadcrumbPage>{jobPosting.job_id}</BreadcrumbPage>
           </BreadcrumbItem>
         </BreadcrumbList>
       </Breadcrumb>
       <div>
-      <h3 className="text-md mb-2 font-semibold text-muted-foreground hover:text-foreground hover-offset-4">
-        
-        <Link className="flex flex-row items-center gap-4" href={`/companies/${jobPosting.company_id}`}>
-        <Avatar alt={jobPosting.companyName} className="w-8 h-8 rounded-full">
-        <AvatarImage src={jobPosting.logo} />
-        <AvatarFallback>{jobPosting.companyName?.charAt(0).toUpperCase()}</AvatarFallback>
-      </Avatar>
-      {jobPosting.companyName}
-      </Link>
-      </h3>
+        <h3 className="text-md mb-2 font-semibold text-muted-foreground hover:text-foreground hover-offset-4">
+
+          <Link className="flex flex-row items-center gap-4" href={`/job-postings?company=${jobPosting.company}`}>
+            <Avatar alt={jobPosting.company} className="w-8 h-8 rounded-full">
+              <AvatarImage src={`https://logo.clearbit.com/${jobPosting.company}.com`} />
+              <AvatarFallback>{jobPosting.company?.charAt(0).toUpperCase()}</AvatarFallback>
+            </Avatar>
+            {jobPosting.company}
+          </Link>
+        </h3>
       </div>
       <h1 data-scroll-title className="text-2xl mb-4 font-semibold decoration-2 leading-normal min-w-0">{jobPosting.title}</h1>
       {keywords && keywords.length > 0 && (
-  <div className="mb-8">
-    <ul className="flex flex-wrap gap-4 gap-y-3">
-      {keywords.map((keyword, index) => {
-        const colors = [
-          { bg: "bg-blue-500/10", text: "text-blue-600", border: "border-blue-600/10" },
-          { bg: "bg-green-500/10", text: "text-green-600", border: "border-green-600/10" },
-          { bg: "bg-yellow-500/10", text: "text-yellow-600", border: "border-yellow-600/10" },
-          { bg: "bg-sky-500/10", text: "text-sky-600", border: "border-sky-600/10" },
-          { bg: "bg-emerald-500/10", text: "text-emerald-600", border: "border-emerald-600/10" },
-          { bg: "bg-red-500/10", text: "text-red-600", border: "border-red-600/10" },
-          { bg: "bg-pink-500/10", text: "text-pink-600", border: "border-pink-600/10" },
-          { bg: "bg-rose-500/10", text: "text-rose-600", border: "border-rose-600/10" },
-          { bg: "bg-purple-500/10", text: "text-purple-600", border: "border-purple-600/10" },
-        ];
+        <div className="mb-8">
+          <ul className="flex flex-wrap gap-4 gap-y-3">
+            {keywords.map((keyword, index) => {
+              const colors = [
+                { bg: "bg-blue-500/10", text: "text-blue-600", border: "border-blue-600/10" },
+                { bg: "bg-green-500/10", text: "text-green-600", border: "border-green-600/10" },
+                { bg: "bg-yellow-500/10", text: "text-yellow-600", border: "border-yellow-600/10" },
+                { bg: "bg-sky-500/10", text: "text-sky-600", border: "border-sky-600/10" },
+                { bg: "bg-emerald-500/10", text: "text-emerald-600", border: "border-emerald-600/10" },
+                { bg: "bg-red-500/10", text: "text-red-600", border: "border-red-600/10" },
+                { bg: "bg-pink-500/10", text: "text-pink-600", border: "border-pink-600/10" },
+                { bg: "bg-rose-500/10", text: "text-rose-600", border: "border-rose-600/10" },
+                { bg: "bg-purple-500/10", text: "text-purple-600", border: "border-purple-600/10" },
+              ];
 
-        const color = colors[index % colors.length]; // Rotate colors based on index
+              const color = colors[index % colors.length]; // Rotate colors based on index
 
-        return (
-          <Badge
-            key={index}
-            variant="outline"
-            className={`${color.bg} ${color.text} rounded-md text-sm sm:text-[13px] font-medium ${color.border}`}
-          >
-            {keyword}
-          </Badge>
-        );
-      })}
-    </ul>
-  </div>
-)}
-      <div className="mb-8 flex flex-wrap gap-4 gap-y-3 text-md font-medium text-neutral-500 items-start">
-        {jobPosting.salary_range_str && (
+              return (
+                <Badge
+                  key={index}
+                  variant="outline"
+                  className={`${color.bg} ${color.text} rounded-md text-sm sm:text-[13px] font-medium ${color.border}`}
+                >
+                  {keyword}
+                </Badge>
+              );
+            })}
+          </ul>
+        </div>
+      )}
+      <div className="mb-8 flex flex-wrap gap-4 gap-y-3 text-md font-medium text-muted-foreground items-start">
+        {jobPosting?.salary ? (
           <div className="flex items-center gap-2">
-            <Zap className="h-[14px] w-[14px] sm:h-4 sm:w-4" />
-            <span>{jobPosting.salary_range_str}</span>
-            </div>
-        )}
-<div className="flex items-center gap-2">
-{(jobPosting?.salary && Number(jobPosting.salary) > 0) || (jobPosting?.salary_max && Number(jobPosting.salary_max) > 0) ? (
-          <div className="flex items-center gap-2">
-            <DollarSign className="h-[14px] w-[14px] sm:h-4 sm:w-4" />
+            <DollarSign className="h-3 w-3 text-muted-foreground" />
             <span>
-              {jobPosting.salary && jobPosting.salary_max
-                ? `$${Number(jobPosting.salary).toLocaleString()} - $${Number(jobPosting.salary_max).toLocaleString()}`
-                : jobPosting.salary_range_str}
+              {jobPosting.salary}
+            </span>
+          </div>
+        ) : jobPosting?.salary_range_str ? (
+          <div className="flex items-center gap-2">
+            <DollarSign className="h-3 w-3 text-muted-foreground" />
+            <span>
+              {jobPosting.salary_range_str}
             </span>
           </div>
         ) : null}
-  <User
-    className={`h-[14px] w-[14px] sm:h-4 sm:w-4 ${
-      jobPosting.applicants === 0 ? "text-green-600" : "text-gray-700"
-    }`}
-  />
-  <span
-    className={jobPosting.applicants === 0 ? "text-green-600" : "text-gray-700"}
-  >
-    {jobPosting.applicants === 0 
-      ? "Be the first applicant" 
-      : `${jobPosting.applicants} applicants`}
-  </span>
-</div>
         <div className="flex items-center gap-2">
-          <MapPin className="h-[14px] w-[14px] sm:h-4 sm:w-4" />
+
+          <User
+            className={`h-[14px] w-[14px] sm:h-4 sm:w-4 ${jobPosting.applicants === 0 ? "text-green-600" : "text-muted-foreground"
+              }`}
+          />
+          <span
+            className={jobPosting.applicants === 0 ? "text-green-600" : "text-muted-foreground"}
+          >
+            {jobPosting.applicants === 0
+              ? "Be the first applicant"
+              : `${jobPosting.applicants} applicants`}
+          </span>
+        </div>
+        <div className="flex items-center gap-2">
+          <MapPin size={16} />
           <span>{jobPosting.location}</span>
         </div>
         <div className="flex items-center gap-2">
-          <Timer className="h-[14px] w-[14px] sm:h-4 sm:w-4" />
-          <span>{formatDistanceToNow(jobPosting.postedDate)}</span>
+          <Timer size={16} />
+          <span>{formatDistanceToNow(jobPosting.created_at, { addSuffix: true })}</span>
         </div>
 
         {jobPosting?.experienceLevel && (
           <>
-        <div className="flex items-center gap-2">
-          <Zap className="h-[14px] w-[14px] sm:h-4 sm:w-4" />
-          <span>{jobPosting.experienceLevel}</span>
-        </div>
-    
-      </>
+            <div className="flex items-center gap-2">
+              <Zap size={16} />
+              <span>{jobPosting.experienceLevel}</span>
+            </div>
+
+          </>
         )}
-        </div>
+      </div>
 
       <div className="flex flex-wrap flex-row gap-4 gap-y-3 mt-4 mb-4">
-        <Link href={`/job-postings?company=${jobPosting.company_id}`}>
-      <NumberButton text={`More Jobs at ${jobPosting.companyName}`} count={companyJobCount} variant="outline" />
-      </Link>
-      <Link className="w-full md:w-auto" href={`${jobPosting.link}`}>
-      
-      <Button className="group md:w-auto text-green-600 bg-green-500/10 border border-green-600/20 hover:bg-green-500/20 hover:text-green-500" >
-      <Briefcase className="-ms-1 me-2 opacity-60" size={16} strokeWidth={2} aria-hidden="true" />
-      Apply on {new URL(jobPosting.link).hostname.split('.').slice(-2, -1)[0]}
-      <ArrowRight 
-        className="-me-1 ms-2 opacity-60 transition-transform group-hover:translate-x-0.5"
-        size={16}
-        strokeWidth={2}
-        aria-hidden="true"
-      />
-    </Button>
-</Link>
-< InsightsButton onClick={handleInghtsClick} />
-{/** 
- *         <ReportPopover />
-        <EnhanceJobPopover jobPosting={jobPosting} />
- * 
-*/}
-
+        <Link href={`/job-postings?company=${jobPosting.company}`}>
+          <NumberButton text={`More Jobs at ${jobPosting.company}`} count={companyJobCount} variant="outline" />
+        </Link>
+        <Link
+          className="w-full md:w-auto"
+          href={`${jobPosting.source_url}`}
+          target="_blank"
+          onClick={handleApplyClick} // Add onClick handler
+        >
+          <Button className="group md:w-auto text-green-600 bg-green-500/10 border border-green-600/20 hover:bg-green-500/20 hover:text-green-500">
+            Apply on {new URL(jobPosting.source_url).hostname.split('.').slice(-2, -1)[0]}
+            <ArrowRight
+              className="-me-1 ms-2 opacity-60 transition-transform group-hover:translate-x-0.5"
+              size={16}
+              strokeWidth={2}
+              aria-hidden="true"
+            />
+          </Button>
+        </Link>
         <Button24 jobId={id} />
-        < CopyButton />
-        </div>
+        <CopyButton />
+        <MagicButton />
 
-{ user && insightsShown && (
-  <>
-        <h3 className="text-md font-semibold mb-3">Quick Insights</h3>
-        <div className="flex flex-wrap gap-2 mb-4">
-                <Badge onClick={handleBadgeClick} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border-foreground/10" variant="secondary">
-                <Sparkle size={14} strokeWidth={2} className="text-muted-foreground" />  
-              </Badge>
-                <Badge onClick={handlePredefinedQuestion} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border-foreground/10" variant="secondary">
-                <User size={14} strokeWidth={2} className="text-muted-foreground mr-2" /> 
+      </div>
 
-                  <p className="text-sm px-1 py-0.5 font-medium text-muted-foreground">
-                    Am I a good fit?
-                    </p>
-                </Badge>
-                <Badge onClick={handleBadgeClick} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border border-foreground/10" variant="secondary">
-                <Briefcase size={14} strokeWidth={2} className="text-muted-foreground mr-2" /> 
+      {user && insightsShown && (
+        <>
+          <h3 className="text-md font-semibold mb-3">Quick Insights</h3>
+          <div className="flex flex-wrap gap-2 mb-4">
+            <Badge onClick={handleBadgeClick} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border-foreground/10" variant="secondary">
+              <Sparkle size={14} strokeWidth={2} className="text-muted-foreground" />
+            </Badge>
+            <Badge onClick={handlePredefinedQuestion} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border-foreground/10" variant="secondary">
+              <User size={14} strokeWidth={2} className="text-muted-foreground mr-2" />
 
-                <p className="text-sm px-1 py-0.5 font-semibold text-muted-foreground">
+              <p className="text-sm px-1 py-0.5 font-medium text-muted-foreground">
+                Am I a good fit?
+              </p>
+            </Badge>
+            <Badge onClick={handleBadgeClick} className="cursor-pointer bg-gradient-to-r from-indigo-500/10 via-purple-500/10 to-pink-500/10 border border-foreground/10" variant="secondary">
+              <Briefcase size={14} strokeWidth={2} className="text-muted-foreground mr-2" />
+
+              <p className="text-sm px-1 py-0.5 font-semibold text-muted-foreground">
                 What should I say in my cover letter?
-                    </p>
-                </Badge>
+              </p>
+            </Badge>
 
+          </div>
+          {llmResponse && (
+            <div className="mb-6 p-4 border rounded-md ">
+              <div dangerouslySetInnerHTML={{ __html: llmResponse }} />
             </div>
-            {llmResponse && (
-                <div className="mb-6 p-4 border rounded-md ">
-                    <div dangerouslySetInnerHTML={{ __html: llmResponse }} />
-                </div>
-            )}
+          )}
 
-            {showAlert && <AlertDemo />}
-            </>
-)}
+          {showAlert && <AlertDemo />}
+        </>
+      )}
       <div className="prose-td code:display-inline-block prose-td code:bg-gray-200 prose-td code:px-2 prose-td code:py-1 prose-td code:rounded-md prose prose-headings:mb-[0.7em] prose-headings:mt-[1.25em] prose-headings:font-semibold prose-headings:tracking-tight prose-h1:text-[32px] prose-h2:text-2xl prose-h3:text-xl prose-h4:text-lg prose-h5:text-base prose-p:mb-4 prose-p:mt-0 prose-p:leading-relaxed prose-p:before:hidden prose-p:after:hidden prose-blockquote:font-normal prose-blockquote:not-italic prose-blockquote:text-neutral-500 prose-blockquote:before:hidden prose-blockquote:after:hidden prose-code:my-0 prose-code:inline-block prose-code:rounded-md prose-code:bg-neutral-100 prose-code:px-2 prose-code:text-[85%] prose-code:font-normal prose-code:leading-relaxed prose-code:text-primary prose-code:before:hidden prose-code:after:hidden prose-pre:mb-4 prose-pre:mt-0 prose-pre:whitespace-pre-wrap prose-pre:rounded-lg prose-pre:bg-neutral-100 prose-pre:px-3 prose-pre:py-3 prose-pre:text-base prose-pre:text-primary prose-ol:mb-4 prose-ol:mt-1 prose-ol:pl-8 marker:prose-ol:text-primary prose-ul:mb-4 prose-ul:mt-1 prose-ul:pl-8 marker:prose-ul:text-primary prose-li:mb-0 prose-li:mt-0.5 prose-li:text-primary first:prose-li:mt-0 prose-table:w-full prose-table:table-auto prose-table:border-collapse prose-th:break-words prose-th:text-center prose-th:font-semibold prose-td:break-words prose-td:px-4 prose-td:py-2 prose-td:text-left prose-img:mx-auto prose-img:my-12 prose-video:my-12 max-w-none overflow-auto text-primary">
-      <Accordion type="single" collapsible className="w-full" defaultValue="item-description">
-        {[
-          { key: 'companyDescription', label: 'Company Description' },
-          { key: 'description', label: 'Job Description' },
-          { key: 'responsibilities', label: 'Responsibilities' },
-          { key: 'requirements', label: 'Requirements' },
-          { key: 'Benefits', label: 'Benefits' },
-          { key: 'MinimumQualifications', label: 'Minimum Requirements' },
-          { key: 'relocation', label: 'Relocation Assistance' },
-          { key: 'EqualOpportunityEmployerInfo', label: 'Equal Opportunity Employer Info' },
-          { key: 'IsRemote', label: 'Remote Work Availability' },
-          { key: 'H1BVisaSponsorship', label: 'H1B Visa Sponsorship' },
-          { key: 'HoursPerWeek', label: 'Hours Per Week' },
-          { key: 'Schedule', label: 'Schedule' },
-          { key: 'NiceToHave', label: 'Nice to Have' },
-          { key: 'raw_description_no_format', label: 'Job Link Description' }
+        <Accordion type="single" collapsible className="w-full" defaultValue="item-description">
+          {[
+            { key: 'companyDescription', label: 'Company Description' },
+            { key: 'description', label: 'Job Description' },
+            { key: 'responsibilities', label: 'Responsibilities' },
+            { key: 'requirements', label: 'Requirements' },
+            { key: 'Benefits', label: 'Benefits' },
+            { key: 'MinimumQualifications', label: 'Minimum Requirements' },
+            { key: 'relocation', label: 'Relocation Assistance' },
+            { key: 'EqualOpportunityEmployerInfo', label: 'Equal Opportunity Employer Info' },
+            { key: 'IsRemote', label: 'Remote Work Availability' },
+            { key: 'H1BVisaSponsorship', label: 'H1B Visa Sponsorship' },
+            { key: 'HoursPerWeek', label: 'Hours Per Week' },
+            { key: 'Schedule', label: 'Schedule' },
+            { key: 'NiceToHave', label: 'Nice to Have' },
+            { key: 'raw_description_no_format', label: 'Job Link Description' }
 
-        ].map(({ key, label }) => (
-          typeof jobPosting[key] === 'string' && jobPosting[key].length > 4 && (
+          ].map(({ key, label }) => (
+            typeof jobPosting[key] === 'string' && jobPosting[key].length > 4 && (
               <AccordionItem className="text-md" key={key} value={`item-${key}`}>
-                  <AccordionTrigger className="text-md font-semibold md:text-lg">{label}</AccordionTrigger>
-                  <AccordionContent className="leading-loose">
-                    <div>
+                <AccordionTrigger className="text-md font-semibold md:text-lg">{label}</AccordionTrigger>
+                <AccordionContent className="leading-loose">
+                  <div>
                     <div
-  dangerouslySetInnerHTML={{
-    __html: stripHTML(decodeHTMLEntities(jobPosting[key])),
-  }}
-/>
+                      dangerouslySetInnerHTML={{
+                        __html: stripHTML(decodeHTMLEntities(jobPosting[key])),
+                      }}
+                    />
 
-                    </div>
-                  </AccordionContent>
+                  </div>
+                </AccordionContent>
               </AccordionItem>
-          )
-        ))}
-      </Accordion>
-</div>
-<div className="flex flex-col space-y-2 mb-4">
-          <p className="text-blue-500 font-medium hover:underline underline-offset-4">
+            )
+          ))}
+        </Accordion>
+      </div>
+      <div className="flex flex-col space-y-2 mb-4">
+        <p className="text-blue-500 font-medium hover:underline underline-offset-4">
           <Link href={`/job-postings?explevel=${encodeURIComponent(jobPosting.experienceLevel)}`}>
-          See more {jobPosting.experienceLevel} jobs
+            See more {jobPosting.experienceLevel} jobs
           </Link>
-          </p>
-          <p className="text-blue-500 font-medium hover:underline underline-offset-4">
+        </p>
+        <p className="text-blue-500 font-medium hover:underline underline-offset-4">
           <Link href={`/job-postings?location=${encodeURIComponent(jobPosting.location.trim())}`}>
             See jobs in {jobPosting.location.trim()}
-            </Link>
-            </p>
+          </Link>
+        </p>
         <p className="text-blue-500 font-medium hover:underline underline-offset-4">
-        <Link href={`/job-postings?title=${encodeURIComponent(jobPosting.title.trim())}`}>
-        See more {jobPosting.title.trim()} jobs
-        </Link>
+          <Link href={`/job-postings?title=${encodeURIComponent(jobPosting.title.trim())}`}>
+            See more {jobPosting.title.trim()} jobs
+          </Link>
 
         </p>
       </div>
       <Suspense fallback={<div>Loading similar jobs...</div>}>
-        <SimilarJobs jobId={id} />
+        <SimilarJobs jobTitle={jobPosting.title} experienceLevel={jobPosting.experienceLevel ?? ""} />
       </Suspense>
 
       <Suspense fallback={<div>Loading similar jobs...</div>}>
-        <CompanySimilarJobs companyId={jobPosting.company_id} />
+        <CompanySimilarJobs company={jobPosting.company} />
       </Suspense>
 
       {relatedPostings?.sameCompanyPostings && relatedPostings.sameCompanyPostings.length > 0 ? (
-  <CollapsibleDemo
-    title={`More Jobs at ${jobPosting?.companyName}`}
-    jobPostings={relatedPostings?.sameCompanyPostings}
-  />
-) : (
-  <div className="text-center text-gray-500 mt-4">
-    No additional job postings from this company.
-  </div>
-)}
+        <CollapsibleDemo
+          title={`More Jobs at ${jobPosting?.company}`}
+          jobPostings={relatedPostings?.sameCompanyPostings}
+        />
+      ) : (
+        <div className="text-center text-gray-500 mt-4">
+          No additional job postings from this company.
+        </div>
+      )}
     </div>
   );
 }
 
-export function ReportPopover() {
+function ReportPopover() {
   return (
     <Popover>
       <PopoverTrigger asChild>
