@@ -8,6 +8,7 @@ import { useAuth } from "@/context/AuthContext"
 import { useToast } from "@/hooks/use-toast"
 import { ToastAction } from "@/components/ui/toast"
 import { Input } from "@/components/ui/input"
+import { Checkbox } from "@/components/ui/checkbox"
 
 import {
   Form,
@@ -30,47 +31,52 @@ import { Textarea } from "@/components/ui/textarea"
 import { useRouter } from "next/navigation"
 
 const SavedSearchSchema = z.object({
+  searchName: z.string().min(1, "Please enter a name for this search"),
   jobTitle: z.string({ required_error: "Please select a job title." }),
   experienceLevel: z.string({ required_error: "Please select an experience level." }),
   location: z.string({ required_error: "Please enter a location." }),
+  notify: z.boolean().default(false),
 })
 
 export function SavedSearchForm({ onSubmit, initialData }) {
-  const { user } = useAuth() // Access user from AuthContext
+  const { user } = useAuth()
   const { toast } = useToast()
   const form = useForm({
     resolver: zodResolver(SavedSearchSchema),
     defaultValues: initialData || {
+      searchName: "",
       jobTitle: "",
       experienceLevel: "",
       location: "",
+      notify: false,
     },
   })
 
   async function handleSubmit(data) {
-    console.log(data)
-    toast({ title: 'Submitting saved search...' })
     if (!user) {
       toast({ title: 'User not authenticated.' })
       return
     }
-    try {
-      toast({ title: 'Saved search created successfully.' })
-      onSubmit(data);
-    } catch (error) {
-      toast({
-        variant: "destructive",
-        title: "Uh oh! Something went wrong.",
-        description: "There was an error creating your saved search.",
-        action: <ToastAction altText="Try again">Try again</ToastAction>,
-      })
-
-    }
+    onSubmit(data);
   }
 
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(handleSubmit)} className="space-y-8 w-full">
+        <FormField
+          control={form.control}
+          name="searchName"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Search Name</FormLabel>
+              <FormControl>
+                <Input placeholder="Give this search a name" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
         <FormField
           control={form.control}
           name="jobTitle"
@@ -101,7 +107,7 @@ export function SavedSearchForm({ onSubmit, initialData }) {
                   <SelectContent>
                     <SelectItem value="internship">Internship</SelectItem>
                     <SelectItem value="entry level">Entry Level</SelectItem>
-                    <SelectItem value="mid level">Mid Level</SelectItem>
+                    <SelectItem value="mid">Mid Level</SelectItem>
                     <SelectItem value="senior">Senior Level</SelectItem>
                     <SelectItem value="manager">Manager</SelectItem>
                     <SelectItem value="director">Director</SelectItem>
@@ -127,7 +133,33 @@ export function SavedSearchForm({ onSubmit, initialData }) {
             </FormItem>
           )}
         />
-        <Button type="submit">Submit</Button>
+
+        <FormField
+          control={form.control}
+          name="notify"
+          render={({ field }) => (
+            <FormItem className="flex flex-row items-start space-x-3 space-y-0 rounded-md border p-4">
+              <FormControl>
+                <Checkbox
+                  checked={field.value}
+                  onCheckedChange={field.onChange}
+                />
+              </FormControl>
+              <div className="space-y-1 leading-none">
+                <FormLabel>
+                  Notify me about new matches
+                </FormLabel>
+                <FormDescription>
+                  Get email notifications when new jobs match this search
+                </FormDescription>
+              </div>
+            </FormItem>
+          )}
+        />
+
+        <Button type="submit">
+          {initialData ? 'Update Search' : 'Save Search'}
+        </Button>
       </form>
     </Form>
   )
