@@ -14,7 +14,7 @@ import { Eye, EyeOff } from "lucide-react";
 import { useState } from "react";
 import { RiFacebookFill, RiGithubFill, RiGoogleFill, RiTwitterXFill } from "@remixicon/react";
 
-export function Button43() {
+function Button43() {
   return (
     <div className="flex flex-wrap gap-2">
       <Button className="flex-1" variant="outline" aria-label="Login with Google" size="icon">
@@ -33,7 +33,7 @@ export function Button43() {
   );
 }
 
-export function Input21() {
+function Input21() {
   return (
     <div className="space-y-2">
       <Label htmlFor="input-21">Email</Label>
@@ -52,7 +52,7 @@ export function Input21() {
   );
 }
 
-export function NotifyMe() {
+function NotifyMe() {
   const [email, setEmail] = useState("");
   const [submitted, setSubmitted] = useState(false);
 
@@ -65,7 +65,7 @@ export function NotifyMe() {
 
   return (
     <div className="space-y-2">
-      <Label htmlFor="notify-me-email">Enter your email to be notified when we're back up:</Label>
+      <Label htmlFor="notify-me-email">Enter your email to be notified when we&apos;re back up:</Label>
       <form onSubmit={handleSubmit} className="flex rounded-lg shadow-sm shadow-black/5">
         <Input
           id="notify-me-email"
@@ -83,31 +83,31 @@ export function NotifyMe() {
           Notify Me
         </button>
       </form>
-      {submitted && <p className="text-sm text-green-500">Thank you! We will notify you when we're back up.</p>}
+      {submitted && <p className="text-sm text-green-500">Thank you! We will notify you when we&apos;re back up.</p>}
     </div>
   );
 }
 
 
-export function AccordionDemo() {
+function AccordionDemo() {
   return (
     <Accordion type="single" collapsible className="w-full">
       <AccordionItem value="item-1">
-        <AccordionTrigger>Where's core?</AccordionTrigger>
+        <AccordionTrigger>Where&apos;s core?</AccordionTrigger>
         <AccordionContent>
-          We're rewriting and renaming it to Junera.
+          We&apos;re rewriting and renaming it to Junera.
         </AccordionContent>
       </AccordionItem>
       <AccordionItem value="item-2">
         <AccordionTrigger>Why rewrite?</AccordionTrigger>
         <AccordionContent>
-          Old code was some sh*t. We're making it better.
+          Old code was some sh*t. We&apos;re making it better.
         </AccordionContent>
       </AccordionItem>
       <AccordionItem value="item-3">
         <AccordionTrigger>What will be different?</AccordionTrigger>
         <AccordionContent>
-          Speed, performance, and quicker shipping of features. 
+          Speed, performance, and quicker shipping of features.
         </AccordionContent>
       </AccordionItem>
     </Accordion>
@@ -115,11 +115,11 @@ export function AccordionDemo() {
 }
 
 
-export function LoginButton() {
+function LoginButton() {
   return <Button className="w-full">Login</Button>;
 }
 
-export function Input31({ ...props }) {
+function Input31({ ...props }) {
   return (
     <div className="group relative w-full">
       <label
