@@ -2,7 +2,140 @@
 import { useEffect } from 'react';
 import { useAuth } from '@/context/AuthContext';
 import { useRouter } from 'next/navigation';
-import { Card, CardTitle, CardDescription } from "@/components/ui/card";
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useForm } from "react-hook-form"
+import { z } from "zod"
+
+import { useToast } from '@/hooks/use-toast';
+import { Button } from "@/components/ui/button"
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form"
+import { Input } from "@/components/ui/input"
+
+const FormSchema = z.object({
+  fullname: z.string().min(2, {
+    message: "Full name must be at least 2 characters.",
+  }),
+  email: z.string().email({
+    message: "Please enter a valid email address.",
+  }),
+  username: z.string().min(2, {
+    message: "Username must be at least 2 characters.",
+  }),
+  password: z.string().min(6, {
+    message: "Password must be at least 6 characters.",
+  }),
+})
+
+function InputForm() {
+  const { toast } = useToast();
+  const form = useForm({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      fullname: "",
+      email: "",
+      username: "",
+      password: "",
+    },
+  })
+
+  function onSubmit(data) {
+    console.log("data: ", data);
+    // Send data to the server
+    fetch(`/api/register?${new URLSearchParams(data).toString()}`, {
+      method: "GET",
+    })
+      .then((res) => res.json())
+      .then((data) => {
+        console.log("data: ", data);
+        if (data.userId) {
+          toast("Account created successfully");
+        } else {
+          toast("Failed to create account");
+        }
+      })
+      .catch((error) => {
+        console.error("Error creating account:", error);
+        toast("Failed to create account");
+      });
+
+    form.reset();
+  }
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="  space-y-6">
+        <FormField
+          control={form.control}
+          name="fullname"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Full Name</FormLabel>
+              <FormControl>
+                <Input placeholder="full name" {...field} />
+              </FormControl>
+              <FormDescription>
+                This is your public display name.
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="email"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Email</FormLabel>
+              <FormControl>
+                <Input placeholder="email" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="username"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Username</FormLabel>
+              <FormControl>
+                <Input placeholder="username" {...field} />
+              </FormControl>
+              <FormDescription>
+                You cannot change this later.
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="password"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Password</FormLabel>
+              <FormControl>
+                <Input type="password" placeholder="password" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <Button type="submit">Submit</Button>
+      </form>
+    </Form>
+  )
+}
+
 
 export default function Login() {
   const { user } = useAuth();
@@ -16,10 +149,8 @@ export default function Login() {
 
   return (
     <div className="flex items-center justify-center min-h-screen p-4 sm:p-8">
-      <main className="flex flex-col gap-8 items-center w-full max-w-md sm:max-w-lg">
-        <Card className="w-full p-4 sm:p-8">
-          <h1 className="text-4xl font-bold text-center mb-4">Login</h1>
-        </Card>
+      <main className="flex flex-col gap-8 w-full max-w-xs sm:max-w-lg">
+        <InputForm />
       </main>
     </div>
   );
