@@ -16,33 +16,43 @@ export default function MatchingJobs({ savedSearches, loading, error }) {
   const [errorJobs, setErrorJobs] = useState(null);
 
   useEffect(() => {
-    
     if (savedSearches && savedSearches?.savedSearches) {
       const fetchJobs = async () => {
         try {
           const allJobs = [];
           for (const search of savedSearches?.savedSearches) {
-            console.log('Search:', search);
-            const searchParams = JSON.parse(search.search_params);
+            const { search_criteria, search_name } = search;
             const query = new URLSearchParams({
-              title: searchParams.jobTitle || '',
-              experienceLevel: searchParams.experienceLevel || '',
-              location: searchParams.location || '',
-              limit: 5 // Limit the number of jobs fetched per search
+              title: search_criteria.title || '',
+              experienceLevel: search_criteria.experienceLevel || '',
+              location: search_criteria.location || '',
+              limit: 20
             }).toString();
 
             const response = await fetch(`/api/job-postings/by-saved-search?${query}`);
+            if (!response.ok) throw new Error('Failed to fetch matching jobs');
+            
             const result = await response.json();
             const jobsWithSearchInfo = result.jobs.map(job => ({
               ...job,
-              matchingString: `Matching: ${searchParams.experienceLevel || 'Any'} • ${searchParams.jobTitle || 'Any'} in ${searchParams.location || 'Any'}`
+              matchedSearch: search_name,
+              matchingCriteria: {
+                title: search_criteria.title,
+                location: search_criteria.location,
+                experienceLevel: search_criteria.experienceLevel
+              }
             }));
             allJobs.push(...jobsWithSearchInfo);
           }
-          const uniqueJobs = Array.from(new Map(allJobs.map(job => [job.id, job])).values());
+          // Remove duplicates and sort by date
+          const uniqueJobs = Array.from(
+            new Map(allJobs.map(job => [job.id, job])).values()
+          ).sort((a, b) => new Date(b.posted_at) - new Date(a.posted_at));
+          
           setJobs(uniqueJobs);
         } catch (error) {
-          setErrorJobs("Error fetching jobs based on saved searches.");
+          setErrorJobs("Error fetching matching jobs");
+          console.error(error);
         } finally {
           setLoadingJobs(false);
         }
@@ -54,28 +64,58 @@ export default function MatchingJobs({ savedSearches, loading, error }) {
     }
   }, [savedSearches]);
 
-  if (loading || loadingJobs) return <div className="space-y-3"><Skeleton className="w-full h-[20px] rounded-full" /><Skeleton className="w-full h-[20px] rounded-full" /><Skeleton className="w-full h-[20px] rounded-full" /></div>;
-  if (error || errorJobs) return <p className="text-red-500">{error || errorJobs}</p>;
-  if (!savedSearches) return <p>No matching jobs found.</p>;
+  if (loading || loadingJobs) {
+    return (
+      <div className="space-y-3">
+        <Skeleton className="w-full h-[20px] rounded-full" />
+        <Skeleton className="w-full h-[20px] rounded-full" />
+        <Skeleton className="w-full h-[20px] rounded-full" />
+      </div>
+    );
+  }
+
+  if (error || errorJobs) {
+    return <p className="text-red-500">{error || errorJobs}</p>;
+  }
+
+  if (!jobs.length) {
+    return <p className="text-muted-foreground text-sm">No matching jobs found for your saved searches.</p>;
+  }
 
   return (
-    <div>
-        <ScrollArea className="h-64">
-      {jobs.map((job) => (
-        <div key={job.id} className="mb-4">
-            <Avatar className="mr-2 h-5 w-5 float-left">
-                <AvatarImage src={job.companyLogo} alt={job.companyName} />
-                <AvatarFallback>{job.companyName[0]}</AvatarFallback>
-            </Avatar>
-        <p className="text-sm text-foreground">{job.companyName}</p>
-          <Link href={`/job-postings/${job.id}`} className="hover:underline">
-          <p className="text-foreground font-medium"><strong className="text-lime-600 font-medium">{job.title}</strong> in {job.location}</p>
+    <ScrollArea className="h-64">
+      <div className="space-y-4">
+        {jobs.map((job) => (
+          <div key={job.id} className="group relative">
+            <div className="flex items-center gap-2">
+              <Avatar className="h-5 w-5">
+                <AvatarImage src={job.companyLogo} alt={job.company} />
+                <AvatarFallback>{job.company?.[0]}</AvatarFallback>
+              </Avatar>
+              <span className="text-sm text-muted-foreground">{job.company}</span>
+            </div>
+            
+            <Link href={`/job-postings/${job.id}`} className="block group-hover:underline">
+              <h3 className="text-foreground font-medium mt-1">
+                <span className="text-lime-600">{job.title}</span>
+                {job.location && <span className="text-muted-foreground"> in {job.location}</span>}
+              </h3>
             </Link>
-          <p className="text-sm text-muted-foreground">{job.experienceLevel}</p>
-          <p className="text-xs text-muted-foreground">{job.matchingString}</p>
-        </div>
-      ))}
-        </ScrollArea>   
-    </div>
+
+            <div className="text-xs text-muted-foreground mt-1">
+              <p>Matching "{job.matchedSearch}"</p>
+              <p className="mt-0.5">
+                {[
+                  job.matchingCriteria.experienceLevel,
+                  job.matchingCriteria.title,
+                  job.matchingCriteria.location
+                ].filter(Boolean).join(' • ')}
+              </p>
+            </div>
+          </div>
+        ))}
+      </div>
+      <ScrollBar />
+    </ScrollArea>
   );
 }
