@@ -8,13 +8,14 @@ import { Skeleton } from '@/components/ui/skeleton';
 
 
 export default function RecentlyAppliedJobs({ jobs, loading, error, router }) {
+  if (!jobs?.appliedJobs) return null;
   if (loading) return <div className="space-y-3"><Skeleton className="w-full h-[20px] rounded-full" /><Skeleton className="w-full h-[20px] rounded-full" /><Skeleton className="w-full h-[20px] rounded-full" /></div>;
   if (error) return <p className="text-red-500">{error}</p>;
-  if (jobs.length === 0) return <p>No recently applied jobs.</p>;
+  if (jobs.appliedJobs.length === 0) return <p>No recently applied jobs.</p>;
 
   return (
     <div>
-      {jobs.slice(0, 3).map((job) => (
+      {jobs.appliedJobs.slice(0, 3).map((job) => (
         <div
           className="mb-2 cursor-pointer"
           key={job.id}
@@ -23,11 +24,11 @@ export default function RecentlyAppliedJobs({ jobs, loading, error, router }) {
           <p className="text-lime-500 hover:underline">{job.company}</p>
           <p className="text-foreground font-medium">{job.title}</p>
           <p className="text-muted-foreground text-xs font-medium">
-            {formatDistanceToNow(new Date(job.postedDate))}
+            {formatDistanceToNow(new Date(job.appliedAt), { addSuffix: true })}
           </p>
         </div>
       ))}
-      {jobs.length > 3 && (
+      {jobs.appliedJobs.length > 3 && (
         <Link href="/job-postings/applied" className="text-lime-500 hover:underline">
           View all applied jobs
         </Link>
