@@ -4,6 +4,7 @@ import Link from 'next/link';
 import { formatDistanceToNow } from 'date-fns';
 import SkeletonCard from './SkeletonCard';
 import { Skeleton } from '@/components/ui/skeleton';
+import { Bell } from 'lucide-react';
 
 export default function SavedSearches({ data, loading, error }) {
     const [jobCounts, setJobCounts] = useState({});
@@ -50,19 +51,23 @@ export default function SavedSearches({ data, loading, error }) {
     return (
         <div className="space-y-4">
             {data.savedSearches.map((search) => {
-                const searchParams = JSON.parse(search.search_params);
-                const jobTitle = searchParams.jobTitle || '';
-                const experienceLevel = searchParams.experienceLevel || '';
-                const location = searchParams.location || '';
-                const createdAt = formatDistanceToNow(new Date(search.created_at)) + ' ago';
+                const { search_name, search_criteria, created_at, notify, is_active } = search;
+                const { title, location, experienceLevel } = search_criteria;
+                const createdAt = formatDistanceToNow(new Date(created_at)) + ' ago';
                 const jobCount = jobCounts[search.id] || '...';
                 const recentJobCount = recentJobCounts[search.id] || '...';
 
                 return (
-                    <div key={search.id} className="">
-                        <Link href={`/job-postings?explevel=${experienceLevel}&location=${location}&title=${jobTitle}`}>
-                            <p className="text-sm font-medium hover:underline">
-                                <strong className="text-lime-600">{jobTitle || 'Any job'}</strong> in {location || 'Anywhere'}
+                    <div key={search.id} className="p-4 border rounded-lg hover:border-lime-500 transition-colors">
+                        <div className="flex items-center justify-between mb-2">
+                            <h3 className="font-medium">{search_name}</h3>
+                            {notify && (
+                                <Bell className="h-4 w-4 text-lime-600" />
+                            )}
+                        </div>
+                        <Link href={`/job-postings?explevel=${experienceLevel}&location=${location}&title=${title}`}>
+                            <p className="text-sm hover:underline">
+                                <strong className="text-lime-600">{title || 'Any job'}</strong> in {location || 'Anywhere'}
                             </p>
                         </Link>
                         <p className="text-foreground text-sm">
@@ -71,6 +76,9 @@ export default function SavedSearches({ data, loading, error }) {
                         <p className="text-muted-foreground text-xs font-medium">
                             {createdAt} • {recentJobCount} new jobs since saved
                         </p>
+                        {!is_active && (
+                            <p className="text-yellow-600 text-xs mt-2">This search is currently inactive</p>
+                        )}
                     </div>
                 );
             })}
